Remove unnecessary (and incorrect) $LOAD_PATH manipulation.
The line in question appends the queue_classic.rb file itself onto the $LOAD_PATH, which is likely not what was intended. Manipulation of the load path is unnecessary with a properly-structured library like this, as the 'require' that loaded queue_classic.rb in the first place wouldn't have succeeded had the necessary directory to load 'queue_classic/okjson' not been in the path.