Feature: Unlock jobs when INT signal is handled.
We no longer are deleting jobs from the queue in the event
of a job exception. If this behavior is desired, it should be defined
in the Worker#handle_failure method.

In the case our worker traps an INT signal and the job is not finished
we will unlock it in the queue.