make it possible to inject a custom worker for QC's rake tasks.
In every project I'm using QueueClassic we end up writing
our own Worker. This is also encouraged in the README.

The process is very smooth but one drawback is that you can't
continue to use the rake tasks, which ship with QC. Or at least
not `rake qc:work` as this would always use the default worker.

I think we should encourage custom workers and make it trivial
to inject them into QC's rake tasks.